post genre

+++ Theatre-system/src/main/java/ru/nsu/theatre/entities/Employee.java
@@ -21,7 +21,7 @@ public class Employee {
     private Integer id;
 
     @Column(name = "type_id")
-    private Integer type_id;
+    private Long type_id;
 
     @Column(name = "date_of_birth")
     private Date date_of_birth;
